+++ consensus/src/test/kotlin/maru/consensus/NextBlockTimestampProviderTest.kt
@@ -12,8 +12,8 @@ import java.time.Clock
 import java.time.Instant
 import java.time.ZoneId
 import kotlin.test.Test
+import maru.config.consensus.ElFork
 import maru.config.consensus.qbft.QbftConsensusConfig
-import maru.config.consensus.qbft.QbftConsensusConfig.Companion.ElFork
 import org.assertj.core.api.Assertions.assertThat
 
 class NextBlockTimestampProviderTest {
