+++ p2p/src/test/kotlin/maru/p2p/QbftForkIdComputerTest.kt
@@ -9,8 +9,8 @@
 package maru.p2p
 
 import kotlin.random.Random
+import maru.config.consensus.ElFork
 import maru.config.consensus.qbft.QbftConsensusConfig
-import maru.config.consensus.qbft.QbftConsensusConfig.Companion.ElFork
 import maru.consensus.ForkId
 import maru.consensus.ForkIdHasher
 import maru.consensus.ForkSpec
