+++ serialization/src/test/kotlin/maru/serialization/ForkSpecSerializerTest.kt
@@ -8,8 +8,8 @@
  */
 package maru.serialization
 
+import maru.config.consensus.ElFork
 import maru.config.consensus.qbft.QbftConsensusConfig
-import maru.config.consensus.qbft.QbftConsensusConfig.Companion.ElFork
 import maru.consensus.ForkSpec
 import maru.core.ext.DataGenerators
 import org.assertj.core.api.Assertions.assertThat
