+++ serialization/src/test/kotlin/maru/serialization/QbftConsensusConfigSerializerTest.kt
@@ -8,8 +8,8 @@
  */
 package maru.serialization
 
+import maru.config.consensus.ElFork
 import maru.config.consensus.qbft.QbftConsensusConfig
-import maru.config.consensus.qbft.QbftConsensusConfig.Companion.ElFork
 import maru.core.ext.DataGenerators
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
